Change sprite pivots
Now set from Custom to Center.
Previously, these were set to the middle of the base of the tile, as opposed to the bottom corner of the base. This forced us to use odd-numbered z-positions in isometric. Now fixed to use even numbers (0, 2, 4, ...). No need to calculate Player offsets either.