Revamp Player Movement and TileManager
TileManager now has a proper, consistent CreateStandableList function. It checks all 4 Tilemaps for whether there is a tile on that coord (for any z-value), then either sets standable to true (if floor or transition) or false (if wall or interactable).
If it turns out that we want to be able to stand on some Interactables, then maybe we can remove the Interactable check later.
Also removes a lot of redundant code.