Revamp isometric floor
I know I just said I finally got rid of manual calculations, but... we're gonna need it again.
The reason why: the way the character renders, it treats the floor just as it would a wall. So if the character's feet are more than halfway behind the wall, it will render behind the wall. This makes it appear as though the character clips through the floor.
How was this resolved before? We simply set the Order in Layer of the Floor to -1. Problem is, this makes the Floor ALWAYS render under the Player. Not inherently bad, BUT...
...I'm hoping to add a stairs functionality, just in case we need it. Was all this unnecessary? Maybe. But it's too late now.

Also it might still have bugs. TBD

I also made the logic simpler and use allPositionsWithin instead of cellBounds-based for-loops.