Add many changes (refer to Description)
Added stair functionality.
Reverted isometric revamp, which effectively gets rid of the whole "-2 here but +2 here and 0 here" business. The Tilemap is just a normal Tilemap, treated normally, starting at 0.
Floor is no longer on -2 (thank goodness).
Changed Transitions/teleports to use a premade Dictionary. Not the most robust thing, but we need a better alternative to TileData based on sprite.
Changed Transparency Sort's z-value to -0.32 because it makes flickering less noticeable. It's the best I can do for now.
Made Player slightly shorter to further prevent flickering.
Added ScanForTile and ScanForTileValue functions, to search a single block at all z-positions. There are built-in functions that do this, but Idk can't really be bothered tbh.