Improve Transition, begin Bed implementation
See TaskManager.cs for Bed (very simple start)

Transition now uses Transition Points (found in the Hierarchy), simply identifies the position of the start and dest "cells" and moves the player accordingly.
Note that Z-positions are crucial for setting these up (and they have to be set up manually, though it's not too difficult since it's all in the Editor).